マイグレーションファイル修正

+++ db/schema.rb
@@ -14,7 +14,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_11_18_121615) do
   create_table "equipment", charset: "utf8", force: :cascade do |t|
     t.string "equipment_name", null: false
     t.string "category", null: false
-    t.float "price", null: false
+    t.integer "price", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
